refactoring for import over staticmethod
